feat: add format scaffold method (#141)
* feat: add format scaffold method

* chore: version and changelog

* feat: call formatScaffoldToProject method
