+++ packages/generate-project/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/generate-project",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "description": "Generator project",
   "files": [
     "lib/"
@@ -14,12 +14,13 @@
     "access": "public"
   },
   "dependencies": {
+    "@iceworks/constant": "^0.1.0",
     "ejs": "^3.1.3",
     "fs-extra": "^8.1.0",
+    "fs-readdir-recursive": "^1.1.0",
     "glob": "^7.1.6",
     "ice-npm-utils": "^2.0.0",
-    "ora": "^4.0.3",
-    "@iceworks/constant": "^0.1.0"
+    "ora": "^4.0.3"
   },
   "gitHead": "fc5b35f95ab4cc24898845916acf598c2f34d576",
   "devDependencies": {
