+++ packages/generate-project/src/formatProject.ts
